requirements.txt
streamlit
pandas
ydata-profiling
great_expectations
presidio-analyzer
presidio-anonymizer
langchain
Faker
phonenumbers
openpyxl
fpdf
matplotlib
google-generativeai
